xio-projects-e2e/Source/server/requirements.txt
appdirs==1.4.3
Django==1.9.5
django-cors-headers==1.3.1
djangorestframework==3.3.3
djangorestframework-jwt==1.9.0
gevent==1.2.1
gevent-websocket==0.10.1
greenlet==0.4.12
grequests==0.3.0
influxdb==4.0.0
lxml==3.6.4
packaging==16.8
psycopg2==2.7.3.1
py==1.4.33
pycrypto==2.6.1
PyJWT==1.4.2
pyparsing==2.2.0
pytest==3.0.7
pytest-django==3.1.2
python-crontab==2.1.1
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
redis==2.10.5
requests==2.12.3
six==1.10.0
